dev-requirements.txt
terminaltables==3.1.0
tabulate==0.7.7
dnspython==1.15.0
coveralls==1.1
pytest-cov==2.2.1
pytest-xdist==1.14
pytest-flake8==0.5
pytest==2.9.2
mypy-lang==0.4.1
